Extract Go analyzer package
